доработала функцию состояния страницы
